Refactor
Use debian:bullseye-slim base image
