+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM golang:1.16-alpine as helpers
+FROM golang:1.19-alpine as helpers
 
 WORKDIR /go/src/helpers
 
@@ -9,52 +9,42 @@ COPY go ./
 RUN CGO_ENABLED=0 go install ./...
 
 
-FROM alpine:20210804 as certificates
+FROM alpine:3.17 as certificates
 
-RUN apk add --no-cache openssl ca-certificates
+RUN apk add --no-cache gnutls-utils
 
-COPY openssl.conf /
+COPY cert.cfg /
+RUN certtool --generate-privkey --outfile tls.key \
+    && certtool --generate-self-signed --load-privkey tls.key --template cert.cfg --outfile tls.crt
 
-# Make a self-signed certificate to satisfy the requirements of the Squid config
-RUN openssl ecparam -name prime256v1 -out ec.param \
- && openssl req -new -newkey ec:ec.param -days 1825 -nodes -x509 -sha384 \
-      -config openssl.conf \
-      -keyout cache-selfsigned.key \
-      -out cache-selfsignedCA.crt \
-      -subj "/O=Veidemann harvester/OU=Veidemann cache/CN=veidemann-harvester"
 
-FROM alpine:20210804
+FROM debian:bullseye-20230109-slim
 
-RUN apk add --no-cache squid tini ca-certificates sudo gettext iptables ip6tables iproute2 && rm -rf /var/cache/apk/*
-
-COPY --from=helpers /go/bin/confighandler /usr/bin/
-COPY --from=helpers /go/bin/storeid /usr/bin/
-COPY --from=helpers /go/bin/loghelper /usr/bin/
-
-COPY --from=certificates --chown=squid:squid /ec.param /ca-certificates/
-COPY --from=certificates --chown=squid:squid /cache-selfsigned.key /ca-certificates/
-COPY --from=certificates --chown=squid:squid /cache-selfsignedCA.crt /ca-certificates/
-
-RUN echo "Cmnd_Alias CMDS = /init-squid.sh, /usr/bin/confighandler" >> /etc/sudoers.d/squid && \
-    echo "squid ALL=(ALL) NOPASSWD: CMDS" >> /etc/sudoers.d/squid && \
-    echo "Defaults:squid !requiretty, !env_reset" >> /etc/sudoers.d/squid && \
-    chmod 440 /etc/sudoers.d/squid
-
-# Use this mount to bring your own certificates
-VOLUME /ca-certificates
-
-VOLUME /var/cache/squid
-
-COPY init-squid.sh /
-COPY docker-entrypoint.sh /
-COPY squid.conf.template /etc/squid/
-COPY squid-balancer.conf.template /etc/squid/
-
-USER squid
-
+ENV TZ=UTC
 ENV SERVICE_NAME="veidemann-cache"
 ENV DNS_SERVERS="8.8.8.8 8.8.4.4"
 
-EXPOSE 3128 3129 4827
+RUN set -eux; \
+    apt-get update; \
+    DEBIAN_FRONTEND=noninteractive apt-get full-upgrade -y; \
+    DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends squid-openssl ca-certificates tzdata; \
+    DEBIAN_FRONTEND=noninteractive apt-get remove --purge --auto-remove -y; \
+    rm -rf /var/lib/apt/lists/*; \
+    sed -i 's/^#http_access allow localnet$/http_access allow localnet/' /etc/squid/conf.d/debian.conf; \
+    echo "# Set max_filedescriptors to avoid using system's RLIMIT_NOFILE. See LP: #1978272" > /etc/squid/conf.d/rock.conf; \
+    echo 'max_filedescriptors 65536' >> /etc/squid/conf.d/rock.conf; \
+    /usr/sbin/squid --version;
 
-ENTRYPOINT [ "/sbin/tini", "--", "/docker-entrypoint.sh" ]
+# Use this mount to bring your own certificates
+VOLUME /ca-certificates
+VOLUME /var/spool/squid
+
+COPY --from=helpers /go/bin/confighandler /go/bin/storeid /go/bin/loghelper /usr/local/sbin/
+COPY --from=certificates --chown=proxy:proxy /tls.key /tls.crt /ca-certificates/
+ADD https://ssl-config.mozilla.org/ffdhe2048.txt /ca-certificates/dhparams.pem
+COPY entrypoint.sh /usr/local/bin/entrypoint.sh
+COPY squid.conf squid.conf.template squid-balancer.conf.template /etc/squid/
+
+EXPOSE 3128
+
+ENTRYPOINT [ "entrypoint.sh"]
