+++ README.md
@@ -1,5 +1,4 @@
 [![License Apache](https://img.shields.io/github/license/nlnwa/veidemann-cache.svg)](https://github.com/nlnwa/veidemann-cache/blob/master/LICENSE)
 [![GitHub release](https://img.shields.io/github/release/nlnwa/veidemann-cache.svg)](https://github.com/nlnwa/veidemann-cache/releases/latest)
-[![Build Status](https://travis-ci.org/nlnwa/veidemann-cache.svg?branch=master)](https://travis-ci.org/nlnwa/veidemann-cache)
 
 # veidemann-cache
