+++ cert.cfg
@@ -1,0 +1,6 @@
+# https://www.gnutls.org/manual/html_node/certtool-Invocation.html
+cn = "veidemann-harvester"
+ca
+signing_key
+cert_signing_key
+encryption_key
