+++ go/.gitignore
@@ -1,3 +1,3 @@
-./confighandler
-./loghelper
-./storeid
+/confighandler
+/loghelper
+/storeid
