--- openssl.conf
@@ -1,14 +1,0 @@
-[ req ]
-distinguished_name     = req_distinguished_name
-x509_extensions         = v3_ca
-
-[ req_distinguished_name ]
-ST = Test State or Province
-O                      = Veidemann
-OU                     = Certificate Authority
-CN                     = Veidemann Web Traffic Recorder
-
-[ v3_ca ]
-basicConstraints = critical, CA:true
-keyUsage=digitalSignature, cRLSign, keyCertSign, keyEncipherment, dataEncipherment
-extendedKeyUsage=serverAuth,clientAuth,anyExtendedKeyUsage
