+++ squid.conf
@@ -1,0 +1,53 @@
+# ----------------------------------------------------------------------------
+# Basic squid configuration
+# ----------------------------------------------------------------------------
+
+# ACL configuration
+# --------------------------
+acl localnet src 0.0.0.1-0.255.255.255  # RFC 1122 "this" network (LAN)
+acl localnet src 10.0.0.0/8             # RFC 1918 local private network (LAN)
+acl localnet src 100.64.0.0/10          # RFC 6598 shared address space (CGN)
+acl localnet src 169.254.0.0/16         # RFC 3927 link-local (directly plugged) machines
+acl localnet src 172.16.0.0/12          # RFC 1918 local private network (LAN)
+acl localnet src 192.168.0.0/16         # RFC 1918 local private network (LAN)
+acl localnet src fc00::/7               # RFC 4193 local private network range
+acl localnet src fe80::/10              # RFC 4291 link-local (directly plugged) machines
+
+# Allow proxing standard web traffic only
+acl SSL_ports port 443
+acl Safe_ports port 80          # http
+acl Safe_ports port 21          # ftp
+acl Safe_ports port 443         # https
+acl Safe_ports port 70          # gopher
+acl Safe_ports port 210         # wais
+acl Safe_ports port 1025-65535  # unregistered ports
+acl Safe_ports port 280         # http-mgmt
+acl Safe_ports port 488         # gss-http
+acl Safe_ports port 591         # filemaker
+acl Safe_ports port 777         # multiling http
+acl CONNECT method CONNECT
+
+# ACLs to target for logging
+acl success_codes http_status 100-199 # informational
+acl success_codes http_status 200-299 # successful transactions
+acl success_codes http_status 300-399 # redirection
+acl failure_codes http_status 400-499 # client error
+acl failure_codes http_status 500-599 # server error
+acl success_hier hier_code HIER_DIRECT
+acl failure_hier hier_code HIER_NONE
+acl failure all-of CONNECT failure_hier
+acl failure all-of !CONNECT failure_codes
+acl success all-of CONNECT success_hier
+acl success all-of !CONNECT success_codes
+
+# ACL operators
+# -------------
+http_access deny !Safe_ports
+http_access deny CONNECT !SSL_ports
+http_access allow localhost manager
+http_access deny manager
+
+coredump_dir /var/spool/squid
+
+include /etc/squid/conf.d/*
+
