+++ squid.conf.template
@@ -1,90 +1,33 @@
-# ----------------------------------------------------------------------------
-# Basic squid configuration
-# ----------------------------------------------------------------------------
-
-# common settings
-# ---------------
 dns_nameservers ${DNS_IP}
-logfile_daemon /usr/bin/loghelper
-
-# cache dir size should be no more than 80% of volume space (/var/cache/squid) size
-cache_dir aufs /var/cache/squid 750 16 256 # Size: 750MB, First level dirs: 16, Second level dirs: 256
-
-coredump_dir /var/cache/squid
-
-forward_max_tries 10
-
-workers 1
 
 # SSL-Bump (see https://wiki.squid-cache.org/ConfigExamples/Intercept/SslBumpExplicit)
-http_port 3128 ssl-bump generate-host-certificates=on dynamic_cert_mem_cache_size=4MB tls-cert=/ca-certificates/cache-selfsignedCA.crt tls-key=/ca-certificates/cache-selfsigned.key tls-dh=prime256v1:/ca-certificates/ec.param cipher=ECDHE-ECDSA-AES256-GCM-SHA384
+http_port 3128 ssl-bump \
+ tls-cert=/ca-certificates/tls.crt \
+ tls-key=/ca-certificates/tls.key \
+ generate-host-certificates=on dynamic_cert_mem_cache_size=4MB \
+ options=NO_SSLv3,NO_TLSv1,NO_TLSv1_1,NO_TLSv1_2,NO_TICKET
 
+sslcrtd_program /usr/lib/squid/security_file_certgen -s /var/spool/squid/ssl_db -M 4MB
 acl step1 at_step SslBump1
 ssl_bump peek step1
 ssl_bump bump all
-
-sslcrtd_program /usr/lib/squid/security_file_certgen -s /var/lib/ssl_db -M 4MB
-
 sslproxy_cert_error allow all
 
-# ACL configuration
-# --------------------------
-acl localnet src 0.0.0.1-0.255.255.255  # RFC 1122 "this" network (LAN)
-acl localnet src 10.0.0.0/8             # RFC 1918 local private network (LAN)
-acl localnet src 100.64.0.0/10          # RFC 6598 shared address space (CGN)
-acl localnet src 169.254.0.0/16         # RFC 3927 link-local (directly plugged) machines
-acl localnet src 172.16.0.0/12          # RFC 1918 local private network (LAN)
-acl localnet src 192.168.0.0/16         # RFC 1918 local private network (LAN)
-acl localnet src fc00::/7               # RFC 4193 local private network range
-acl localnet src fe80::/10              # RFC 4291 link-local (directly plugged) machines
+maximum_object_size 64 MB
+offline_mode on
 
-# Allow proxing standard web traffic only
-acl SSL_ports port 443
-acl Safe_ports port 80          # http
-acl Safe_ports port 21          # ftp
-acl Safe_ports port 443         # https
-acl Safe_ports port 70          # gopher
-acl Safe_ports port 210         # wais
-acl Safe_ports port 1025-65535  # unregistered ports
-acl Safe_ports port 280         # http-mgmt
-acl Safe_ports port 488         # gss-http
-acl Safe_ports port 591         # filemaker
-acl Safe_ports port 777         # multiling http
-
-# ACLs to target for logging
-acl success_codes http_status 100-199 # informational
-acl success_codes http_status 200-299 # successful transactions
-acl success_codes http_status 300-399 # redirection
-acl failure_codes http_status 400-499 # client error
-acl failure_codes http_status 500-599 # server error
-acl success_hier hier_code HIER_DIRECT
-acl failure_hier hier_code HIER_NONE
-acl failure all-of CONNECT failure_hier
-acl failure all-of !CONNECT failure_codes
-acl success all-of CONNECT success_hier
-acl success all-of !CONNECT success_codes
-
-# ACL operators
-# -------------
-http_access deny !Safe_ports
-http_access deny CONNECT !SSL_ports
-http_access allow localhost manager
-http_access deny manager
-
-# Set PID file to a place the default squid user can write to
-pid_filename /var/run/squid/squid.pid
+# Refresh patterns
+refresh_pattern .               4320    100%    4320 store-stale override-expire override-lastmod ignore-private ignore-reload ignore-no-store
 
 # Logging configuration
 # ---------------------
-debug_options ALL,0
+debug_options ALL,1
 log_mime_hdrs off
-access_log stdio:/dev/stdout
-cache_log stdio:/dev/stderr
-cache_store_log stdio:/dev/stderr
+logfile_daemon /usr/local/sbin/loghelper
 
 # Cache settings
 # --------------------
-store_id_program /usr/bin/storeid -t 20
+store_id_program /usr/local/sbin/storeid -t 20
 store_id_extras "%{veidemann_jeid}>h"
 store_id_children 20 startup=5 idle=5 concurrency=20
 
@@ -92,17 +35,14 @@ cache_miss_revalidate off
 negative_ttl 100 day
 request_entities on
 
-maximum_object_size 64 MB
-
-offline_mode on
-
-# Refresh patterns
-refresh_pattern .               4320    100%    4320 store-stale override-expire override-lastmod ignore-private ignore-reload ignore-no-store
-
 # Manipulate headers
 # ----------------------
 request_header_access veidemann_eid deny all
 request_header_access veidemann_jeid deny all
 
-# visible_hostname ensures that X-Cache header always contains 'from veidemann_cache' after HIT/MISS
+# Ensure that X-Cache header always contains 'from veidemann_cache' after HIT/MISS
 visible_hostname veidemann_cache
+
+# Example cache dir config
+# Cache size should be no more than 80% of volume space (/var/spool/squid) size
+# cache_dir aufs /var/spool/squid 750 16 256 # Size: 750MB, First level dirs: 16, Second level dirs: 256
